% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_target}
\alias{add_target}
\title{Augment observed attainment data with "future observations" as targets}
\usage{
add_target(., lvl, tgt, by_yr)
}
\arguments{
\item{.}{The dataframe of observed attainment to which the target should be added}

\item{lvl}{The target attainment level}

\item{tgt}{The target share as a fraction}

\item{by_yr}{The year by which the target is reached}
}
\description{
Function \code{add_target} is a convenience wrapper to simplify the
specification of (e.g. SDG) targets for educational expansions.
}

