% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicedproj.R
\name{edproj}
\alias{edproj}
\title{Run the educational attainment projection}
\usage{
edproj(x)
}
\arguments{
\item{x}{A dataframe in the correct input format (as generated by \code{wic_prepare}, for example)}
}
\description{
Function \code{edproj} is the main front-end function
to run all steps involved in the projection. Currently
it is mandatory to include a target, as both trend and
target-attaining trajectories are estimated.
}

