% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{wic_prepare}
\alias{wic_prepare}
\title{Title}
\usage{
wic_prepare(x)
}
\arguments{
\item{df}{Input data of (reconstructed) attainment. Assumed to contain
variables: \code{country} (factor), region (factor), gender (factor), age_group (factor), year (numeric), level
(factor, increasing order, including residual category),}
}
\description{
Title
}

