% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicedproj.R
\docType{package}
\name{wicedproj}
\alias{wicedproj}
\alias{wicedproj-package}
\title{\code{wicedproj}: A package implementing the
Education Projection Model of the Wittgenstein Centre for Demography and Global Human Capital.}
\description{
\code{wicedproj}: A package implementing the
Education Projection Model of the Wittgenstein Centre for Demography and Global Human Capital.
}

