% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{wicedproj_options}
\alias{wicedproj_options}
\title{Set or get options for oostypo package}
\usage{
wicedproj_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for oostypo package
}
\section{Supported options}{

The following options are supported:
\itemize{
  \item{\code{testing}}{(\code{boolean};1) Whether to run only a single chain for a few iterations
  for testing purposes.}
}
}

